# [1533] Optimise providers show