@tvararu 2019-05-24
![](https://github.trello.services/images/mini-trello-icon.png) [Optimise organisation page on Publish](https://trello.com/c/cX699Bbz/1533-optimise-organisation-page-on-publish)